Move 995 helper functions to shared
