+++ src/applications/appeals/995/reducers/contestableIssues.js
@@ -1,5 +1,4 @@
-import { getEligibleContestableIssues } from '../utils/helpers';
-
+import { getEligibleContestableIssues } from '../../shared/utils/issues';
 import contestableIssuesReducer from '../../shared/reducers/contestableIssues';
 
 export default contestableIssuesReducer(getEligibleContestableIssues);
